+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Enhancement] Provide `Karafka::Messages::Message#tombstone?` to be able to quickly check if a message is a tombstone message.
 - [Enhancement] Provide more flexible API for Recurring Tasks topics reconfiguration.
 - [Enhancement] Remove no longer needed Rails connection releaser.
+- [Fix] Fix a case where there would be a way to define multiple subscription groups for same topic with different consumer.
 
 ## 2.4.9 (2024-08-23)
 - **[Feature]** Provide Kafka based Recurring (Cron) Tasks.